+++ x-pack/legacy/plugins/integrations_manager/public/hooks/use_core.tsx
